Move discussion out of `types` into `lib`
It contains more than just types now. It would also be useful to move the discussion reducer into this file to make use of React's context API (which can't be used while the reducer code lives in the component file).
