+++ dotcom-rendering/fixtures/manual/discussion-no-top-comments.ts
@@ -1,4 +1,4 @@
-import type { GetDiscussionSuccess } from '../../src/types/discussion';
+import type { GetDiscussionSuccess } from '../../src/lib/discussion';
 
 export const discussionNoTopComments = {
 	status: 'ok',
