+++ dotcom-rendering/fixtures/manual/topPicks.ts
@@ -1,4 +1,4 @@
-import type { GetDiscussionSuccess } from '../../src/types/discussion';
+import type { GetDiscussionSuccess } from '../../src/lib/discussion';
 
 export const topPicks = {
 	status: 'ok',
