+++ dotcom-rendering/src/components/Discussion/Dropdown.test.tsx
@@ -1,6 +1,6 @@
 import '@testing-library/jest-dom/extend-expect';
 import { fireEvent, render, screen } from '@testing-library/react';
-import type { DropdownOptionType } from '../../types/discussion';
+import type { DropdownOptionType } from '../../lib/discussion';
 import { Dropdown } from './Dropdown';
 
 const threadOptions: [
