+++ dotcom-rendering/src/components/Discussion/Filters.stories.tsx
@@ -1,7 +1,7 @@
 import { ArticleDesign, ArticleDisplay, Pillar } from '@guardian/libs';
 import { useState } from 'react';
 import { splitTheme } from '../../../.storybook/decorators/splitThemeDecorator';
-import type { FilterOptions } from '../../types/discussion';
+import type { FilterOptions } from '../../lib/discussion';
 import { Filters } from './Filters';
 
 export default { title: 'Discussion/Filters' };
