+++ dotcom-rendering/src/components/Discussion/Pagination.stories.tsx
@@ -1,7 +1,7 @@
 import { ArticleDesign, ArticleDisplay, Pillar } from '@guardian/libs';
 import { useState } from 'react';
 import { splitTheme } from '../../../.storybook/decorators/splitThemeDecorator';
-import type { FilterOptions } from '../../types/discussion';
+import type { FilterOptions } from '../../lib/discussion';
 import { Pagination } from './Pagination';
 
 const articleFormat: ArticleFormat = {
