+++ dotcom-rendering/src/components/Discussion/TopPick.stories.tsx
@@ -1,8 +1,8 @@
 import { css } from '@emotion/react';
 import { ArticleDesign, ArticleDisplay, Pillar } from '@guardian/libs';
 import { splitTheme } from '../../../.storybook/decorators/splitThemeDecorator';
+import type { CommentType, SignedInUser } from '../../lib/discussion';
 import { ok } from '../../lib/result';
-import type { CommentType, SignedInUser } from '../../types/discussion';
 import { TopPick } from './TopPick';
 
 export default { component: TopPick, title: 'Discussion/TopPick' };
