+++ dotcom-rendering/src/components/Discussion/TopPicks.tsx
@@ -1,6 +1,6 @@
 import { css } from '@emotion/react';
 import { from, until } from '@guardian/source-foundations';
-import type { CommentType, SignedInUser } from '../../types/discussion';
+import type { CommentType, SignedInUser } from '../../lib/discussion';
 import { TopPick } from './TopPick';
 
 type Props = {
