+++ dotcom-rendering/src/components/DiscussionMeta.importable.tsx
@@ -1,9 +1,9 @@
 import { joinUrl } from '@guardian/libs';
+import type { GetDiscussionSuccess, UserProfile } from '../lib/discussion';
 import { getOptionsHeadersWithOkta } from '../lib/identity';
 import { useApi } from '../lib/useApi';
 import { useAuthStatus } from '../lib/useAuthStatus';
 import { useCommentCount } from '../lib/useCommentCount';
-import type { GetDiscussionSuccess, UserProfile } from '../types/discussion';
 import { SignedInAs } from './SignedInAs';
 
 type Props = {
