+++ dotcom-rendering/src/components/HeaderTopBarMyAccount.tsx
@@ -2,6 +2,7 @@ import { css } from '@emotion/react';
 import { joinUrl } from '@guardian/libs';
 import { from, palette, textSans } from '@guardian/source-foundations';
 import { useEffect, useState } from 'react';
+import type { UserProfile } from '../lib/discussion';
 import { getZIndex } from '../lib/getZIndex';
 import type {
 	AuthStatus,
@@ -18,7 +19,6 @@ import { nestedOphanComponents } from '../lib/ophan-helpers';
 import { useApi } from '../lib/useApi';
 import { useBraze } from '../lib/useBraze';
 import ProfileIcon from '../static/icons/profile.svg';
-import type { UserProfile } from '../types/discussion';
 import type { RenderingTarget } from '../types/renderingTarget';
 import { useConfig } from './ConfigContext';
 import type { DropdownLinkType } from './Dropdown';
